Cast to fix MSVC build

+++ src/vsp_aero/Viewer/glviewer.C
@@ -363,7 +363,7 @@ void GL_VIEWER::LoadInitialData(char *name)
     if (pathsep)
     {
         sprintf( file_name_no_path, "%s", pathsep + 1 );
-        sprintf( path, "%.*s", pathsep - file_name, file_name );
+        sprintf( path, "%.*s", ( int ) ( pathsep - file_name ), file_name );
     }
     else
     {
